Implement numpy support
All of the actual mathematical tools can be used from the numpy module
instead of the math module, and they will work transparently for arrays
as well as floats. This version works both with and without numpy; and
it should behave as before if numpy is not installed, or if it is
installed and the inputs are floats. The only issues involved are
related to bounds checking and other inequalities.

Bounds checking is done with the utility function in_bounds which checks
if ALL elements satisfy the conditions.

For calculating the zone letters and numbers, only one element is
used: the first one. It is assumed to be caller’s responsibility that
all points lie in the same zone. Forcing the zone may be useful for
numpy users, therefore added a force_zone_letter argument to
from_latlon, to make the zone fully forceable.

The only other caveat is that the south/north check in from_latlon may
cause problems if the input latitude has mixed signs. This is therefore
checked and may raise a ValueError.

Tests do assume that numpy is installed, however.