Fix forcing zones around equator and add force_northern in from_latlon
Logic in from_latlon has been changed to be similar to to_latlon,
including adding a force_northern option to mirror northern option in
to_latlon. This resolves the issue with northing coordinate incorrectly
being set when pasing in -ve or +ve latitude when forcing northern or
southern zone respectively.

Closes #35 - alternative fix for issue.