Fix issue near anti-meridian when forcing zones
This resolves an issue with zones near the anti-meridian when converting
points on the opposing side due to forcing zones. The cause was due to
the difference in angle being incorrectly calculated e.g. difference
between a longitude of 179 and -179 should be 2, not -358.

This change also ensures that longitudes returned are ≥ -180 and < 180